Added Event folder and moved appropriate items
