+++ SafeguardDotNet/A2A/ISafeguardA2AContext.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Security;
+using OneIdentity.SafeguardDotNet.Event;
 
 namespace OneIdentity.SafeguardDotNet.A2A
 {
