+++ SafeguardDotNet/SafeguardDotNet.csproj
@@ -24,8 +24,4 @@
     <PackageReference Include="Serilog" Version="2.7.1" />
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="Event\" />
-  </ItemGroup>
-
 </Project>
